+++ src/Schema/Directives/FieldDirective.php
@@ -3,6 +3,7 @@
 namespace Nuwave\Lighthouse\Schema\Directives;
 
 use GraphQL\Type\Definition\ResolveInfo;
+use Nuwave\Lighthouse\Schema\RootType;
 use Nuwave\Lighthouse\Schema\Values\FieldValue;
 use Nuwave\Lighthouse\Support\Contracts\FieldResolver;
 use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
@@ -38,7 +39,7 @@ GRAPHQL;
 
         $namespacedClassName = $this->namespaceClassName(
             $className,
-            $fieldValue->defaultNamespacesForParent()
+            RootType::defaultNamespaces($fieldValue->getParentName())
         );
 
         $resolver = Utils::constructResolver($namespacedClassName, $methodName);