+++ src/Subscriptions/SubscriptionResolverProvider.php
@@ -7,6 +7,7 @@ use GraphQL\Type\Definition\ResolveInfo;
 use Illuminate\Support\Str;
 use Nuwave\Lighthouse\Exceptions\DefinitionException;
 use Nuwave\Lighthouse\Schema\AST\ASTHelper;
+use Nuwave\Lighthouse\Schema\RootType;
 use Nuwave\Lighthouse\Schema\Types\GraphQLSubscription;
 use Nuwave\Lighthouse\Schema\Values\FieldValue;
 use Nuwave\Lighthouse\Subscriptions\Directives\SubscriptionDirective;
@@ -47,7 +48,7 @@ class SubscriptionResolverProvider implements ProvidesSubscriptionResolver
 
         $className = Utils::namespaceClassname(
             $className,
-            $fieldValue->defaultNamespacesForParent(),
+            RootType::defaultNamespaces($fieldValue->getParentName()),
             function (string $class): bool {
                 return is_subclass_of($class, GraphQLSubscription::class);
             }